Merge of 905 branch.. The vs c++ api passes 883 testsuite cases (when running in script) The xe c++ api passes 888 testsuite cases (when running in script)
The vs c++ api passes 886 testsuite cases (when running in unit tester)
The xe c++ api passes 884 testsuite cases (when running in unit tester)

Observe that this merge breaks the linux workings. The main change is in rrCompiler where tcc is called. The old 'linux' code should be reused and a #if defined(__linux) can be used.




git-svn-id: https://roadrunnerlib.googlecode.com/svn/trunk@261 c0027639-9190-0b57-01eb-f09a03790dae