Getting threading model implemented
Added job queue to the roadrunner thread class. One job is mainly one RoadRunner instance..
Added job Mutex and condition variable in order to keep thread and main thread to access the same data at the same time..

The main time for roadrunner in a typical 'run' is spent on loading, compiling the model. The simulation itself seem to be pretty fast.
One should be able to re-use the generated model (DLL) and just set the data. Need more investigation..

Removed attempts to keep the two C-API's in one file.. it may be too confusing.

git-svn-id: https://roadrunnerlib.googlecode.com/svn/branches/create_cm_api@636 c0027639-9190-0b57-01eb-f09a03790dae