Adding reset functions
Removed the reset() in the simulation group and replaced it with three
similar functions in their own group (called "reset").

The new reset() function resets variables to their initial values. The
new resetAll() resets the variables to their initial values and
parameters to original values. The new resetToOrigin() resets the
variables to the ORIGINAL initial values and the parameters to original
values.