Fix poco timeout issue.
I can't figure out how this ever worked:  the call to 'getLogger' locked a static member variable... and then called a subfunction which *also* called 'getLogger', which then can't lock the already-locked object.  This works happily on Windows, but times out in the latest gcc on Linux.  My solution is to only lock the static member variable in the block of 'getLogger' that creates a new logger if one does not exist, since the subcalls just need a link to the newly-created object.

The program still doesn't work on Linux, but it doesn't work in a new place, at least.