Throw if trying to add a rule for a floating species.
Also had to fix a bunch of tests that accidentally created invalid models.  We chose to throw for boundary conditions but not for constant because the user cannot actually set 'constant' when creating any element, but must do it separately if they don't want the default.
(Also fixed the wording in our other throws, since they can be called both for setting assignment rules and rate rules.)