Add the ability to set the max output rows in a simulation.
As part of addressing the unintuitive behavior in roadrunner's 'simulate' and 'gillespie' algorithms, we here add the ability to set the maximum number of output rows by the user, instead of always simply taking a default.

Also, the default was changed from 40,000 to 100,000.

Finally, in a gillespie variable step size simulation, the number of steps is now ignored if the duration is specified, and if the duration is set to zero, it takes the set number of steps instead.