Fix the non-static rrplugins libraries so they compile.
For now, the only way I could find to make this work was to make literally everything always link the static libraries.  I have a feeling that there's some other option somewhere I could find to let it work a different way, but since this works, I'm going with it for now.

Also, some of the classes needed to not be set export the way they were.