Fix plugin installation.
The format I ended up going with was to put *everything* that has to do with the rrplugins into site-packages/rrplugins.  This includes the roadrunner C API dll, which has roadrunner linked statically to it.  The upshot is that 'rrplugins' doesn't rely on the existence of the 'roadrunner' package at all--it has its own version in a dll.

It's possible that there will be problems trying to port objects from one dll to another, but this approach does work with at least some of the tested plugins.

(Also fixed up a few python bugs/warnings in the interface.)