Fix bug where simulate results were not constant.
The default option for copy_result should have been 'true', but it defaulted to 'true' because when I was revamping things, I grabbed the wrong option.

I considered removing the option to not copy the result at all, but it turned out that some 'power users' actually do use the option as a speedup, and just know that when they do, they run the risk of invalidating the memory in their old results.