Fix for #725
The check to see if we hit the max number of output points wasn't making sure that 'time' was in the first column.  It also had an aborted error message.

Also, another check would trigger only in situations where it wasn't helpful to warn about it, so I took out that warning.