Convert DoubleMatrix properties to TelluriumData properties
When accessing the DoubleMatrix elements from Python (the Hessian and the Covariance Matrix), the old code that converted a DoubleMatrix to numpy matrix was failing.  I don't understand why this is, but since the TelluriumData conversion works (because it copies things element-by-element), I converted the Hessian and the CovarianceMatrix to be TelluriumData properties.

Other changes:
* Fix up the LM and NM python examples (both of which now run!)
* Update formatting for add_noise.cpp
* Only warn when 'setData' doesn't have column headers.
* Report more information on exceptions in the plugin manager.