Fix the Monte Carlo plugin
Mostly, this involved changing the MC plugin so that it used the direct API instead of going through the C API.  The latter was problematic because while it was possible to register the various plugins themselves, all the properties *also* needed to be registered, and we couldn't do this automatically without changing how things were linked, which I was loathe to do.  This way, we use the standard Plugin interface, which, since the Monte Carlo plugin is itself a plugin, it clearly has access to the normal Plugin interface.

Other changes include:

* Fix the calculation of confidence limits.
* Fix the Python example.
* Add a MC test.
* Put the plugin in its own namespace like the others.
* Allow the creation of StringList properties with StringLists (gasp).