Refactor and simplify solver creation.
Prior to this commit, the mechanism for creating a Solver involved a complicated mechanism whereby a SolverRegistrar and a SolverMgr interacted with a SolverFactory to register solvers. The problems associated with this design include: 1) much code duplication, 2) unnecessarily complicated and 3) this design requires both static and non-static versions of the same methods to be implemented in all solver, which is bulky and horrible.

In this commit we have reduced the complexity of the registration system by refactoring the content of "Registrar" types into an pure abstract interface. The Solver class, base to all integrators, steady state solvers (and soon sensitivities) implements this interface but leaves the methods abstract, passing the responsibility for implementing these methods on to individual solver implementations. Now, instead of creating a SolverRegistrar to register a particular solver, we can just instantiate the Solver with default construction (i.e. without a model instance). This mechanism elegantly resolves all three issues above.