Give up and make the plugins tests Windows-only.
After some research, I discovered that if you create tests for the auto2000 plugin in the main 'develop' branch, you get the same sort of linking errors I was getting in the plugins-develop-merge branch, plus some other fun ones.  The upshot is that the tests do run on Windows, and all old functionality (such as it was) should be maintained going forward.

There are still problems, obviously, but I'm going to claim that it's OK to go ahead and merge this in, since the problems were *revealed* by the new tests and were not *created* by the new code.