Fix setting/getting initial values (almost)
This branch re-creates the old issue-711 branch, and picks up where it left off.  It turns out that 'not setting initial boundary species' is a systemic problem that was simply forgotten in the initial implementation of roadrunner, and fixing it necessitated the creation of a wide variety of low-level functions to create a parallel ecosystem (as it were) for boundary species alongside the floating species.  Basically, a lot of the old code assumed that all initial values for species would always be floating, which is very simply very wrong.

There's still some problems with state saving, and I have yet to test whether resetting the model works properly (and it probably doesn't).  But this at least implements the ability to set and get the initial values for boundary species.

A couple other changes:
* Include more header files so that VS can actually find them when it searches for stuff.
* Some error message fixes (capitalization).
* Be more specific about why we won't load a previous or later version of a roadrunner binary file.