Fix bug: model generator not copied correctly.
The number of global parameters was being assigned to the number of initial boundary species (!)

In tracking down this bug, I used the model printing to cout functions, which turned out to be printing doubles in hex, for some reason, so switched to printing them out normally.  Also, added printing out more things that hadn't been printed before.