Set up consistent rules about what gets reset, and when.
This addresses issue #710 and #811, and should be completely working for compartments, but not totally for other elements yet.  The basic idea is that we are much more careful about resetting elements and checking whether they're set by rate rules.  At this point, if COMPARTMENT is requested, all compartments are actually reset, same with BOUNDARY and GLOBAL_PARAMETER.  Similarly, but more complicatedly, we also do the same for RATE, for *all* elements instead of just global parameters.

Global parameters are still a bit weird with the conserved moiety bit they have, but I left the old code mostly intact for that, so that should still work.  Hopefully.

Still need to implement changing the compartment fix for all the other types as well, which is a lot of copying and pasting (or, perhaps, figuring out how to make it a function).