Apply compartment fix to everything else.
Now that everything works for compartments, apply that elsewhere.  Managed to get this into a single function with passed-in functors, which was a lot more trouble than it should have been, but worked out in the end.

For some reason, species initial assignments are computed from other initial states, while compartment/parameter initial assignments are computed from other current states.  This makes them behave differently even when everything else is equal.  I think I'm punting on this, and will just file a separate bug report for it.