Fix references to the beginnings of empty arrays.
Noticed that a test didn't test what it said it did ('internalOneStep'); when I changed so it did, found a separate bug that apparently only triggers if the new test is run.  This turned out to be having a pointer to an empty array, i.e. &array[0].  Fixed this the same way we did elsewhere in the code, and found one other place the same fix needed to be applied.

Also fixed a typo in another test name.