Add option to allow requested time vector output to simulate.
Also, change the signature of 'simulate' to require a SimulateOptions instead of a Dictionary, since if we didn't get a SimulateOptions, we throw away the Dictionary!  The docs are also updated, as they could not have ever worked in their previous form, too.  I also made SimulateOptions not a Dictionary at all, since literally none of its dictionary settings were ever checked for any reason.  The docs (again) claimed that you could set the integrator here, but this also was a pack of lies.

For the substance of the change, we now allow a new 'times' option in SimulateOptions that can contain a vector of times, and check it with a new 'initialize' function.  Then we use a new 'getNext' function inside 'simulate' instead of calculating from hstep.  The options object will then hand out 'next' values that are either calculated from start/duration or from the time vector.

The change in Roadrunner from simulating to hstep (previously) to simulating to next - prev (now) caused the LM optimizer to take a wildly different turn and ended up with ridiculously different results.  It's probably a bad idea to leave those numeric tests in at all, but for now I corrected them to fit what I got--we'll see if we actually get those same results on other platforms.

Finally, the 'copy_result' option was being used inconsistently from Python--it was being set and gotten by calling 'setItem' but retrieved by querying the member variable, which was never synced with the setItem calls.  Since the object is no longer a dictionary at all, and therefore doesn't have 'setItem', I changed it to be consistent, and always use the boolean member variable.  (It's an odd design choice to store a python-only setting in a C++ object, but anyway.)