Undeprecate everything.
Lots of things have been marked 'deprecated' in roadrunner for the last 7 years, but are still used.  In some cases, this just turned out to be a mistake, and the functions are actually useful.  In other cases, nobody ever went through and removed their use internally, so at this point, there's no good reason to deprecate them, either.  This just marks everything as non-deprecated.