Enable negative start times.
I thought this would be much more difficult than it turned out to be, because our current 'start' time is not 'time to start collecting data' but 'the time that the model is currently in', which is quite different!

The trickiest part of this turns out to be events, which had t0=0.0 hard-coded in quite a number of places.  Still need to test more, but many things now work with negative start times--events can fire pre-0.0, even.  So hey!