Extend negative starts to gillespie.
Also, add a bunch of tests.

Also also, it turns out that events that fire at t0 with delay don't interact well with non-zero start times (nor do events that fire at t0 with 'time' in the trigger).  But that's a separate issue and doesn't involve only negative start times; I'll deal with that separately.