Fix FROM_SCRATCH_7.
For some reason, this one test was causing hundreds of memory leak errors.  When I changed the roadrunner model to being a local object, they all went away!  It's mysterious, but at this point I'm not going to complain.

Also made the unique_ptr objects in the RoadRunner impl protected, to ensure they weren't being messed with elsewhere.