More memory fixes, courtesy of Valgrind.
In one mockup test, volumes were being queried that didn't exist.

In the forward sensitivity solver, we were creating a matrix from Sundials, then changing our reference of the size of that matrix, then destroying the matrix based on the new reference size instead of the original.  This led to a memory leak.