Fix SWIG wrapping of various getX functions.
This turned out to be an issue with SWIG changing, as far as I can tell--it used to translate a Python array into 'int len, type* values', and now it translates it into 'size_t len, type* values', so our old 'int/type' prototype functions no longer matched.  Happily, that meant that it was a relatively straightforward fix.