Add the ability to set and get the name and id of SBML models.
There was already a roadrunner getModelName function and an executable model getModelName function (the former of which was explicitly excluded from the python interface), but this will allow users to set and get both the name and the id at will.  I also changed the model getModelName function so it uses the id if the name isn't set.

As part of testing this functionality, noticed that invalid SBML fails due to an invalid read, so fixed that to properly throw a more informative error.  In the course of doing that, found a libsbml bug, which I submitted a fix for (https://github.com/sbmlteam/libsbml/pull/169)