Fix two more tests.
The tests themselves needed updating here:  one relied on arccot(0) being different from arccot(-0), which is dumb, and another listed the expected variables in the wrong order.

We had passed 959 before because we special-cased it, but I fixed that one, too, because again, that's silly.

There are comments in the roadrunner code about how the arccot test is dumb, but nobody ever complained to the SBML team!  Complain about dumb tests, people!