Change debug message to error.
If we're unable to draw from a truncated distribution after a bunch of tries, we give up and return the midpoint instead.  This needs to be mentioned as an error message, so that users will realize they've set their truncation limits too small, and/or that they need to re-code their model to not use that exact call.

Also made messages more specific.