Fix the sequential times vs. start/end/points bug.
The python version of the simulate functions starts with the built-in 'opt' object, which persists from simulation to simulation.  This probably has always been a problem, but it was exacerbated by the new 'times' option, which is incompatible with the 'start/end/points' options.  So, add a new 'reset' function to the Options object, and use it when calling 'simulate'.