Reorg to attempt proper compilation.
It turned out that rrplugins_core and rrplugins_c_api were including each other.  Moved a few things out of the c api and into core so that it would be strictly hierarchical.

Not sure if this will actually solve the mac/linux problem, but at least it's neater now?