More char* -> const char* catches.
The deal here is that if you call a function with a string literal (i.e. 'foo("hello")') the argument can't be a char*, only a const char* (in the strict warning sense).  I think the idea is that a char* can be changed, which you can't do with a string literal.