Properly raise AttributeError with __getattr__
Addresses #901:  hasattr was failing because __getattr__ wasn't raising an AttributeError on failure.

Not sure if this ever worked, despite the report.  Other objects in roadrunner had correct __getattr__functions that properly failed, but 3 of the 4 defined here did not.  Affected the Solver, the SteadyStateSolver, and the Integrator.  (The RoadRunner object was fine.)