Fix https://github.com/sys-bio/roadrunner/issues/905: correctly pass error message from libsbml
The problem here turned out to be in part a bug in libsbml that would return NULL instead of the error in question if the error was an XML error and not an SBML Error.  I worked with the libsbml team to fix this, but in the meantime, this changed code should work both with the current SBML and the fixed version (the fixed version will have more information, though).