The full Jacobian was sometimes being calculated incorrectly.
The problem was that in order to set conserved moiety values, the initial values of the species were being used as a sort of 'scratch pad' to calculate the unscaled species elasticities.  Unfortunately, this relied on the compartment volumes not having changed since their initial values.

The solution was simple:  at the same time we overwrite and later restore the initial species values, we overwrite and later restore the initial compartment volumes.