Allow a wider variety of selections.
Roadrunner defines an almost dizzying array of selection record types, but hasn't handled all of them.  This adds the ability to handle more types, and adds tests for them.

Of particular note, this adds the ability to handle the FLOATING_CONCENTRATION_RATE type, even when the species amount and the compartment volume are both changing (denoted by the string "[S1]'".  Other changes include:

* The ability to get rate rule values the same way you get other values in models.
* 'suplied' -> 'supplied'
* The addition of the COMPARTMENT_RATE record type.
* Lots more selection record tests.