Make 'getId' (and thus 'keys()') return full list.
There were many more options for 'getValue' that were not listed in 'getIds':  this adds them in to the list, and makes sure they can all be calculated with 'getValue'.  Elasticity and control coefficients can't be calculated for elements set by assignment rules because we can't perturb them.  This was filed as a new bug, and those strings are left out of the getIds list.

As part of this effort, a few other bugs were noticed and quashed:
* The boundary species concentration rate was not being retrieved, and is now.
* Requesting 'eigen(x)' in C++ always returned zero.  It now returns the real part if the imaginary part is zero, and throws an error otherwise.
* Elasticity calculation now throws an error if asked to calculate the elasticity of a conserved moiety.
* Calculating a control coefficient no longer changes the model state.  (Or, rather, it does, but we reset it afterwards.)