Move and rename python tests; remove old code.
This commit does a lot of moving things around and deleting old code.  In the end, all the python tests are moved to test/python/ and then installed to roadrunner/tests/.  'roadrunner.runTests()' works, but no effort has been made to actually ensure that the python tests are run in Azure or anywhere else, yet.

The bits that are removed mostly were broken anyway, with the exception of the source/testing/ directory, which worked, but didn't create anything that actually *did* anything.  They were confusing, and nobody should ever have used them, so out they go.

The C rrplugins examples didn't work because of the consistent use of 'bool' in the plugins API, which I didn't find any quick fixes for.  Instead, we'll just have people use C++.

The source/llvm_testing stuff had already been replicated in the test/* directories, and were not being compiled.