Numerious problems were fixed in save state and the original save state test suite. The main issue was where the binary stream containing the object file was not saved with saveState. This problem doesn't matter until you try to save the state of a copy of a roadrunner model. Copying a roadrunner model also uses save state, so by not transferring the binary stream containing the compiled model caused segfault when trying to access the stream to convert the binary stream to string (in MCJit, LLJit was fine).
This exposed some problems with the save state test suite. The main issue is that the generic string "save-state-test.rr" was used for a file path. Where exactly c++ decided to save this I don't know, but I suspect in home or somewhere similar. This means that the first time the save state tests are run, this file is generated and then reused for every test! Tests must be independent and therefore this is a bug. Fixing this problem revealed several tests that loadState before having first saved the state. These have been disabled.