Switch to probably-debugged version of function.
There were two ModelGeneratorContext constructors; one with a SBML Document, and one with an SBML string.  Originally, I just deleted the version with the string, since I now wanted to use the version with the document.  But it seems that the one with the document was not debugged, and caused problems, since the one with the string was actually used in roadrunner.  So, this switches to using the version with the string, modified to use a document instead.