Revert pipeline and update stoch test.
Now that I finally have the information I need, I can revert the azure pipeline to do what it used to do before I started messing with it.  Then I can update the stochastic test suite to only run a handful of tests on manylinux-azure, which is super slow for unknown reasons.