This replaces getElementBySId where possible with faster lookups.
If we're actually looking for a species, just call 'getSpecies'; if we're looking for anything with mathematical meaning, use the new 'getElementWithMathematicalMeaning' static function in the roadrunner class.  (I tried putting this in rrUtils, but that seems to be formulated for a slightly different interface, so I put it in RoadRunner instead.  If there's a better place for it, I'm happy to put it there instead.  I suppose one place might be libsbml itself, with the advantage that it could deal with plugins with mathematical meaning better, though there aren't a lot of them.)

Also, MockSBase needed updating, as some functions changed their signatures in the latest libsbml.

It would still be nice if we kept a hash of IDs somewhere, and it would mean we could check uniqueness a little faster elsewhere (where we still use getElementBySId), but that's a bigger project.