Connect the 'saveState' version to rr's version.
Recreates old branch that no longer merges.

It's possible that this might be simultaneously too small a change and too large a change:  if the version number changes without the saveState data changing, we needlessly refuse to load files that would have loaded.  Conversely, if a 'patch' version change changes the saveState data, we might not catch it.

I think this is a reasonable compromise, as it's difficult to remember to change the dataVersionNumber otherwise--it's been stuck on '1' ever since it was introduced, which has included a few saveState changes.  But if we need to change it to major+minor+patch, that'd also be OK by me.