When calculating ucc, don't fail if we try to set something already set.
To calculate ucc's, we use the current model state as a sort of scratch pad for calculations, and then have to reset everything again after we're done.  However, some of those set values might already be set due to assignment rules.  So I had to implement a whole 'strict' system for the setters so that they could continue even when asked to do something dumb, because figuring things out manually would be even more work and rigamarole.