Step 1 towards building universal binaries on a mac: brew install llvm.
Reports are that brew install llvm works on a mac, not only for the normal mac, but for the M1 as well.  Hopefully this means that llvm is installed on brew as a universal binary.  Under that assumption, this first change just checks to see if we can figure out how to use the brew installed version of llvm instead of our own version.