Fix for #779:  remove warnings.
LLVM will emit a *lot* of warnings if you let it.  We wrap llvm includes in #pragmas that disable the warning messages they give.  With the LLVM revamp, some new files evaded the old #pragmas, so this puts them back in.

Also fixed a warning that was a real problem, since it had been obscured by all the llvm warnings (an '&' instead of an '&&')