Fix the rate labels when getting *all* the rates.
Also managed to discover where the problem was with the volume/concentration rates:  when d[S1]/dt is set via a rate rule and its compartment's dC/dt is also set by a rate rule, the rate of change for the amount needs to be:

[S1]*dC/dt + C*d[S1]/dt.

However, the math was reversed, so it was instead:

C*dC/dt + [S1]*d[S1]/dt

This mostly fixes the VolumeConcentrationRates tag in the SBML test suite, but I didn't turn that on yet because one of the tests has an unrelated problem with event triggers.

Also fixed a couple error messages that I found along the way.

Still haven't actually tested the original cause of this problem: 'getRatesOfChangeNamedArray' in Python.  Let's hope these changes fix that.