Fix bug when 'time' is involved in a reaction rate.
When calculating the unscaled elasticities, the routine uses a 'scratch pad' of the initial values of the model so that any conserved moieties can be adjusted.  This caused problems, as setting the initial values reset the model, setting 'time' to -inf (for 'events that fire at t0' reasons).  We fix this by saving the time, then just setting it back to that every time we call 'setInitialValue'.

Also, clean up a few things here and there, including llvm warning messages, typos, and a couple tests.