A variety of cleanups and error catching code.
Changed the new functions to match the old functions a little better and be more efficient, i.e. naming functions with 'V' instead of without to match other similar functions, and copying values all at once instead of twice.

Also, I'm not sure how the old tests worked, since it seems to me like they shouldn't have:  the 'steadystate.xml' model only has one dependent and one independent species if you turn on conserved moieties, which the test hadn't done.

In the course of doing that, I found some crashes, which turned out to be because if x is an empty vector, &x[0] is undefined.  We had a lot of code that never checked this, which in the course of checking in Python, caused the kernel to crash.  Added wrappers for all of them.

The C++ tests turned out to not be running; they had to be added to RoadRunnerAPITestsWithLLJit.cpp and RoadRunnerAPITestsWithMCJit.cpp before that would happen.  This revealed other errors, which I can fix in a subsequent commit if Claire is already gone ;-)