Revamp!  Now it doesn't matter if conserved moiety analysis is on.
After talking with Herbert, everything is now changed so that asking about dependent/independent anything will now (nigh) always give you the constrained version (i.e. check if there are conserved cycles).  The *model* still performs its local analysis based on whether or not conserved moiety analysis is on or not, but everything else now assumes you're analyzing the model in its platonic state, instead of depending on it being analyzed in a particular way.

This also meant that the behavior of the Python bindings was changed, so that r.getNumDepFloatingSpecies and r.getNumIndFloatingSpecies now behave the same as r.getNumberOfDependentSpecies and r.getNumberOfIndependentSpecies.  Which now both exist, instead of only one existing, for some reason.

Also changed some functions to not duplicate code, so now the get___NamedArray functions call get___V for the data.