Fix!
The problem turned out to be that the local copy of the stoichiometry matrix was being incorrectly copied, so it only worked in some cases.  However, this was the only place the LLVM's 'getStoichiometryMatrix' function was being called, so I removed it entirely (especially since it was broken).  The remaining 'getStoichiometry(species, rxn)' function is fine, so I swapped that in for the old