Clear off space as we build.
It *might* be that the rrplugins distributions were larger than they needed to be because we didn't clear out the build/ directory before building the rrplugins wheels.  But more importantly, the ubuntu builds were dying because the disk was filling up, so deleting the build directory should help.

Also updated numpy requirement to 1.26, since that's what the build system is using.