Fix for rateOf() species in reactions.
I was returning a reference to a local object, which gets destroyed.  The solution is that when we are asked for one, we create one and stick it in the list.  This means the function can't be const any more, but this seems to not actually matter, other than needing to put const_casts everywhere.

Also, check for recursive calls to rateOf(sp) in reactions as well as those in rate rules.